fix: put cache inside AbstractController, instead of BlocksController
